Extract the structured permission response contract
